Make the sounddevice module optional.
Only imported if -o (output) is not used e.g.

Relevance, there's no Debian package for sounddevice, whereas
scipy and numpy are available in Debian.

Even if sounddevice module is installed (e.g. via pip), if
an actual sound device isn't present (in WSL or ssh session)
this also prevents a module import that would fail.

No need to put the user through that if they're just using -o
or equivilent by way of the api.